requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
environs==11.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
mysqlclient==2.2.4
packaging==24.0
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
